ESC-657 migrate BusinessEntityController to FormHelpers trait
